+++ _integrations/logcontext/nrlogrusplugin/nrlogrusplugin_test.go
@@ -126,7 +126,7 @@ func TestLogNoContext(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestLogNoContext",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestLogNoContext",
 		"timestamp":   float64(1417136460000),
 	})
 }
@@ -140,7 +140,7 @@ func TestLogNoTxn(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestLogNoTxn",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestLogNoTxn",
 		"timestamp":   float64(1417136460000),
 	})
 }
@@ -161,7 +161,7 @@ func TestLogDistributedTracingDisabled(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestLogDistributedTracingDisabled",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestLogDistributedTracingDisabled",
 		"timestamp":   float64(1417136460000),
 	})
 }
@@ -190,7 +190,7 @@ func TestLogSampledFalse(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestLogSampledFalse",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestLogSampledFalse",
 		"timestamp":   float64(1417136460000),
 		"trace.id":    "d9466896a525ccbf",
 	})
@@ -220,7 +220,7 @@ func TestLogSampledTrue(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestLogSampledTrue",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestLogSampledTrue",
 		"span.id":     "bcfb32e050b264b8",
 		"timestamp":   float64(1417136460000),
 		"trace.id":    "d9466896a525ccbf",
@@ -254,7 +254,7 @@ func TestEntryUsedTwice(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestEntryUsedTwice",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestEntryUsedTwice",
 		"span.id":     "bcfb32e050b264b8",
 		"timestamp":   float64(1417136460000),
 		"trace.id":    "d9466896a525ccbf",
@@ -278,7 +278,7 @@ func TestEntryUsedTwice(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World! Again!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestEntryUsedTwice",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestEntryUsedTwice",
 		"timestamp":   float64(1417136460000),
 	})
 }
@@ -301,7 +301,7 @@ func TestEntryError(t *testing.T) {
 		// Since the err field on the Entry is private we cannot record it.
 		//"logrus_error": `can not add field "func"`,
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestEntryError",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestEntryError",
 		"timestamp":   float64(1417136460000),
 	})
 }
@@ -322,7 +322,7 @@ func TestWithCustomField(t *testing.T) {
 		"line.number": matchAnything,
 		"log.level":   "info",
 		"message":     "Hello World!",
-		"method.name": "github.com/newrelic/go-agent/_integrations/log-plugins/nrlogrusplugin.TestWithCustomField",
+		"method.name": "github.com/newrelic/go-agent/_integrations/logcontext/nrlogrusplugin.TestWithCustomField",
 		"timestamp":   float64(1417136460000),
 		"zip":         "zap",
 	})