+++ CHANGELOG.md
@@ -50,7 +50,10 @@ support instrumentation of publishers and subscribers.
 
   * [NATS Example](https://github.com/newrelic/go-agent/go-agent/blob/master/_integrations/nrnats/examples/main.go)
   * [NATS Streaming Example](https://github.com/newrelic/go-agent/go-agent/blob/master/_integrations/nrstan/examples/main.go)
-
+  
+* Enables ability to migrate to [Configurable Security Policies (CSP)](https://docs.newrelic.com/docs/agents/manage-apm-agents/configuration/enable-configurable-security-policies) on a per agent basis for accounts already using [High Security Mode (HSM)](https://docs.newrelic.com/docs/agents/manage-apm-agents/configuration/high-security-mode). 
+  * Previously, if CSP was configured for an account, New Relic would not allow an agent to connect without the `security_policies_token`. This led to agents not being able to connect during the period between when CSP was enabled for an account and when each agent is configured with the correct token.
+  * With this change, when both HSM and CSP are enabled for an account, an agent (this version or later) can successfully connect with either `high_security: true` or the appropriate `security_policies_token` configured - allowing the agent to continue to connect after CSP is configured on the account but before the appropriate `security_policies_token` is configured for each agent. 
 
 ## 2.11.0